requirements.txt
DateTime==6.0
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.61.1
fpdf2==2.8.7
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
zope.interface==8.2
